feat(sbb-seat-reservation-refactor-scoped-parts): only one scoped element
