+++ src/elements-experimental/seat-reservation/seat-reservation/scoped-components.ts
@@ -1,4 +1,1 @@
-export * from './scoped-components/coach-border.js';
-export * from './scoped-components/coach-wrapper.js';
-export * from './scoped-components/table-cell.js';
-export * from './scoped-components/seat-reservation-graphical-element.js';
+export * from './scoped-components/scoped-element.js';
