--- src/elements-experimental/seat-reservation/seat-reservation/scoped-components/coach-border.scss
@@ -1,6 +1,0 @@
-.coach-border {
-  position: absolute;
-  inset-inline-start: var(--sbb-coach-border-inset-inline-start);
-  inset-block-start: var(--sbb-coach-border-inset-block-start);
-  z-index: 0;
-}
