--- src/elements-experimental/seat-reservation/seat-reservation/scoped-components/coach-wrapper.scss
@@ -1,4 +1,0 @@
-.coach-wrapper {
-  height: var(--sbb-coach-wrapper-height);
-  width: var(--sbb-coach-wrapper-width);
-}
