+++ src/elements-experimental/seat-reservation/seat-reservation/scoped-components/scoped-element.scss
@@ -1,0 +1,23 @@
+:host {
+  display: block;
+
+  .coach-border,
+  .coach-wrapper,
+  .graphical-element {
+    width: var(--sbb-scoped-width);
+    height: var(--sbb-scoped-height);
+    inset-block-start: var(--sbb-scoped-inset-block-start);
+    inset-inline-start: var(--sbb-scoped-inset-inline-start);
+    z-index: var(--sbb-scoped-z-index);
+  }
+
+  .coach-border,
+  .graphical-element ::slotted(*),
+  .graphical-element {
+    position: absolute;
+  }
+
+  .coach-border {
+    z-index: 0;
+  }
+}
