--- src/elements-experimental/seat-reservation/seat-reservation/scoped-components/seat-reservation-graphical-element.scss
@@ -1,12 +1,0 @@
-.sbb-seat-reservation__graphical-element ::slotted(*) {
-  position: absolute;
-}
-
-.sbb-seat-reservation__graphical-element {
-  position: absolute;
-  inset-block-start: var(--sbb-seat-reservation-graphical-element-inset-block-start);
-  inset-inline-start: var(--sbb-seat-reservation-graphical-element-inset-inline-start);
-  width: var(--sbb-seat-reservation-graphical-element-width);
-  height: var(--sbb-seat-reservation-graphical-element-height);
-  z-index: var(--sbb-seat-reservation-graphical-element-z-index);
-}
